Added metadata titles for each page on the web app so that the tab says where the user is
